Moves link.cpp/h to common from redalert.
